# openapi/openapi.json
{
  "components": {
    "responses": {
      "Problem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "Error in RFC 9457 format."
      }
    },
    "schemas": {
      "Diagnostic": {
        "description": "Typst compilation diagnostic.",
        "properties": {
          "column": {
            "minimum": 0,
            "type": "integer"
          },
          "file": {
            "description": "Path relative to the template folder.",
            "type": "string"
          },
          "hints": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "line": {
            "minimum": 0,
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
      "DocumentMetadata": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "oneOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "items": {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                "maxItems": 20,
                "minItems": 1,
                "type": "array"
              }
            ]
          },
          "date": {
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}$",
            "type": "string"
          },
          "keywords": {
            "items": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "type": "array"
          },
          "subject": {
            "maxLength": 2000,
            "type": "string"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ErrorCode": {
        "description": "Machine-readable error code (FR-019).",
        "enum": [
          "template-not-found",
          "invalid-json",
          "invalid-parameter",
          "unsupported-media-type",
          "validation-failed",
          "payload-too-large",
          "template-invalid",
          "render-failed",
          "render-timeout",
          "overloaded"
        ],
        "type": "string"
      },
...
      "Problem": {
        "description": "RFC 9457 error body.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": "string"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "type": "array"
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "templateId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri",
            "type": "string"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/Violation"
            },
            "type": "array"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object"
      },
      "RenderRequest": {
        "additionalProperties": false,
        "description": "Generic shape; the exact shape of `data` and `layout` is given by the template's schema\n(`GET /templates/{templateId}/schema`).",
        "properties": {
          "data": {
            "additionalProperties": {},
            "description": "Business data.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "layout": {
            "additionalProperties": {},
            "description": "Layout parameters; missing properties take their `default`.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "metadata": {
            "$ref": "#/components/schemas/DocumentMetadata",
            "description": "PDF metadata, all optional. Without `author`, the service's default author\n(`INKPDF_DEFAULT_AUTHOR`, `inkpdf` by default); without `title`, the template's title,\nor failing that its name. `date` in `YYYY-MM-DD` format."
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "Violation": {
        "description": "A request body's violation of a template's schema.",
        "properties": {
          "message": {
            "type": "string"
          },
          "path": {
            "description": "JSON Pointer into the request body.",
            "type": "string"
          },
          "schemaPath": {
            "description": "JSON Pointer into the schema.",
            "type": "string"
          }
        },
        "required": [
          "path",
          "schemaPath",
          "message"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Internal service (private VPC, no authentication). Templates are hot-loaded from a mounted volume; the caller sends only JSON (`data` + `layout`), validated against the template's JSON Schema before generation.",
    "license": {
      "name": "TODO(LICENSE) — MIT OR Apache-2.0 proposed"
    },
    "title": "inkpdf",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/templates/{templateId}/render": {
      "post": {
        "description": "Validates the body against the template's schema (after applying the `default` values of\n`layout`), then returns the PDF. No generation is attempted if validation fails.",
        "operationId": "renderTemplate",
        "parameters": [
          {
            "description": "Template identifier (its folder name).",
            "example": "sample",
            "in": "path",
            "name": "templateId",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$",
              "type": "string"
            }
          },
          {
            "description": "`true`: respond as a download (`Content-Disposition: attachment`) instead of\ninline display. Values: `true`, `false`, `1`, `0`.",
            "in": "query",
            "name": "download",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "description": "File name (no path; `.pdf` appended). Default: `<templateId>.pdf`.",
            "in": "query",
            "name": "filename",
            "required": false,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "contentMediaType": "application/pdf",
                  "type": "string"
                }
              }
            },
            "description": "Generated document.",
            "headers": {
              "Content-Disposition": {
                "description": "`inline` (display) or `attachment` (download, `download=true`), with `filename` and `filename*` (UTF-8).",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Body does not match the template's schema."
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          },
          "504": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Generate a PDF",
        "tags": [
          "render"
        ]
      }
    },
    "/templates/{templateId}/schema": {
      "get": {
        "description": "Usable directly by a validator or a form generator.",
        "operationId": "getTemplateSchema",
        "parameters": [
          {
            "description": "Template identifier (its folder name).",
            "example": "sample",
            "in": "path",
            "name": "templateId",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/schema+json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "JSON Schema (draft 2020-12)."
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "summary": "Raw JSON Schema of the template input",
        "tags": [
          "templates"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
}